refactor(/media/backend): Rename modules + fix one docstring
